Created filemanager class from actions.py  and moved code
